+++ src/main.c
@@ -40,7 +40,7 @@ struct Config config = {
     "OS",       // os_prefix
     "Kernel",   // kernel_prefix
     "Desktop",  // desktop_prefix
-    "Theme",    // theme_prefix
+    "Theme",    // gtk_theme_prefix
     "Shell",    // shell_prefix
     "Login",    // login_shell_prefix
     "Terminal", // term_prefix
@@ -397,7 +397,7 @@ int main(int argc, char **argv) {
         {"os", config.os_prefix, os},
         {"kernel", config.kernel_prefix, kernel},
         {"desktop", config.desktop_prefix, desktop},
-        {"theme",config.theme_prefix, theme},
+        {"gtk_theme", config.gtk_theme_prefix, gtk_theme},
         {"shell", config.shell_prefix, shell},
         {"login_shell", config.login_shell_prefix, login_shell},
         {"term", config.term_prefix, term},
@@ -425,7 +425,7 @@ int main(int argc, char **argv) {
             "os",
             "kernel",
             "desktop",
-            "theme",
+            "gtk_theme",
             "shell",
             "term",
             "packages",